Attempt retries on communication failures on Add.
This takes a look at the number of idle connections in our pool and
allows that many retries.  This is a bit flawed in that we don't really
manage the pool ourselves, and this code would be better weaved together
to ensure that we only try the retries on connections that are being
reused, rather than freshly generated.

If we created a facade around the connections to wrap operations that
would allow us much greater control over retries to allow stale
connections to be dealt with more cleanly.