Attempt retries on communication failures on Add.
This notices that there are idle connections in the pool, and therefore
infers that we could well be given one.  This is used to determine that
it want's to allow a retry if we get an EOF when we try to send.

If we created a facade around the connections to wrap operations that
would allow us much greater control over retries to allow stale
connections to be dealt with more cleanly.